LPS-180450 Renaming packages
