+++ modules/util/portal-tools-db-partition-virtual-instance-migrator/src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/util/DatabaseUtil.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.util;
+package com.liferay.portal.tools.db.partition.virtual.instance.migrator.util;
 
 import com.liferay.portal.kernel.dao.db.DBInspector;
 import com.liferay.portal.kernel.version.Version;
-import com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.release.Release;
+import com.liferay.portal.tools.db.partition.virtual.instance.migrator.Release;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
