+++ modules/util/portal-tools-db-partition-virtual-instance-migrator/src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/util/DatabaseUtilTest.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.util;
+package com.liferay.portal.tools.db.partition.virtual.instance.migrator.util;
 
 import com.liferay.portal.kernel.version.Version;
-import com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.release.Release;
+import com.liferay.portal.tools.db.partition.virtual.instance.migrator.Release;
 
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
