+++ modules/util/portal-tools-db-partition-virtual-instance-migrator/src/test/java/com/liferay/portal/tools/db/partition/virtual/instance/migrator/util/ValidatorTest.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.util;
+package com.liferay.portal.tools.db.partition.virtual.instance.migrator.util;
 
 import com.liferay.portal.kernel.version.Version;
-import com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.recorder.Recorder;
-import com.liferay.portal.tools.db.partition.virtual.instance.migrator.internal.release.Release;
+import com.liferay.portal.tools.db.partition.virtual.instance.migrator.Recorder;
+import com.liferay.portal.tools.db.partition.virtual.instance.migrator.Release;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
